+++ nextgisweb/feature_attachment/nodepkg/attachment-table/index.ts
@@ -1,0 +1,4 @@
+/** @entrypoint */
+import { AttachmentTable } from "./AttachmentTable";
+
+export default AttachmentTable;