+++ nextgisweb/feature_attachment/nodepkg/image-thumbnail/index.ts
@@ -1,0 +1,5 @@
+import { ImageThumbnail } from "./ImageThumbnail";
+
+export default ImageThumbnail;
+
+export { type ImageThumbnailProps } from "./ImageThumbnail";